refactor: build-logic 내부 구현을 위한 패키지 구조 변경
* refactor(build-logic): 내부 구현을 위한 `internal` 패키지 추가 및 관련 파일 이동
* refactor(build-logic): `localProperty` 유틸리티를 `external` 패키지로 이동
* refactor(network): `BASE_URL` 설정 로직 단순화
* chore(data): 불필요한 의존성 제거
