+++ Prezel/build-logic/convention/src/main/java/com/team/prezel/buildlogic/convention/internal/ProjectExtensions.kt
@@ -1,9 +1,9 @@
-package com.team.prezel.buildlogic.convention
+package com.team.prezel.buildlogic.convention.internal
 
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalog
 import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.kotlin.dsl.getByType
 
-val Project.libs
+internal val Project.libs
     get(): VersionCatalog = extensions.getByType<VersionCatalogsExtension>().named("libs")
