+++ Prezel/build-logic/convention/src/main/java/com/team/prezel/buildlogic/convention/plugin/AndroidApplicationConventionPlugin.kt
@@ -1,7 +1,7 @@
 package com.team.prezel.buildlogic.convention.plugin
 
 import com.android.build.api.dsl.ApplicationExtension
-import com.team.prezel.buildlogic.convention.configureKotlinAndroid
+import com.team.prezel.buildlogic.convention.internal.configureKotlinAndroid
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
