+++ Prezel/build-logic/convention/src/main/java/com/team/prezel/buildlogic/convention/plugin/AndroidLibraryConventionPlugin.kt
@@ -1,8 +1,8 @@
 package com.team.prezel.buildlogic.convention.plugin
 
 import com.android.build.api.dsl.LibraryExtension
-import com.team.prezel.buildlogic.convention.configureKotlinAndroid
-import com.team.prezel.buildlogic.convention.libs
+import com.team.prezel.buildlogic.convention.internal.configureKotlinAndroid
+import com.team.prezel.buildlogic.convention.internal.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
