+++ Prezel/build-logic/convention/src/main/java/com/team/prezel/buildlogic/convention/plugin/HiltConventionPlugin.kt
@@ -1,6 +1,6 @@
 package com.team.prezel.buildlogic.convention.plugin
 
-import com.team.prezel.buildlogic.convention.libs
+import com.team.prezel.buildlogic.convention.internal.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
