+++ Prezel/core/data/build.gradle.kts
@@ -10,6 +10,5 @@ android {
 
 dependencies {
     implementation(projects.core.network)
-
     implementation(libs.kotlinx.coroutines.core)
 }
