+++ Prezel/core/network/build.gradle.kts
@@ -1,5 +1,5 @@
 import com.android.build.api.variant.BuildConfigField
-import com.team.prezel.buildlogic.convention.localProperty
+import com.team.prezel.buildlogic.convention.external.localProperty
 
 plugins {
     alias(libs.plugins.prezel.android.library)
@@ -25,7 +25,6 @@ dependencies {
     implementation(libs.kotlinx.serialization.json)
     implementation(libs.timber)
 
-    // Ktorfit
     implementation(libs.ktorfit.lib)
     ksp(libs.ktorfit.ksp)
 
@@ -35,22 +34,11 @@ dependencies {
 androidComponents {
     onVariants { variant ->
         val buildConfigFields = variant.buildConfigFields ?: return@onVariants
-        val isRelease = variant.buildType == "release"
-        // DEBUG_RELEASE_BASE_URL 또는 RELEASE_BASE_URL
-        val key = "${variant.buildType!!.uppercase()}_BASE_URL"
-
-        val urlProvider = if (isRelease) {
-            localProperty(key).map {
-                it.ifEmpty { throw GradleException("$key is required for release builds") }
-            }
-        } else {
-            localProperty(key).orElse("http://10.0.2.2")
-        }
-
+        val key = "${variant.buildType?.uppercase()}_BASE_URL"
         buildConfigFields.put(
             "BASE_URL",
-            urlProvider.map { value ->
-                BuildConfigField("String", """"$value"""", null)
+            localProperty(key).map { value ->
+                BuildConfigField("String", "\"$value\"", null)
             },
         )
     }
